# web/public/slick.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "slick.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/auth/isAuthorized/{CompanyName}/{ProjectName}/{Permission}": {
      "get": {
        "operationId": "IsAuthorized",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/slickqaIsAuthorizedResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "CompanyName",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "ProjectName",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "Permission",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "LoginWithCredentials",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/slickqaLoginResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/slickqaPlainUserLoginRequest"
            }
          }
        ],
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/login-with-token/{Token}": {
      "get": {
        "operationId": "LoginWithToken",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/slickqaLoginResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "Token",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/users/by-company/{CompanyName}": {
      "get": {
        "operationId": "GetUsersForCompany",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/slickqaUsersQueryResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "CompanyName",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/by-project/{CompanyName}/{ProjectName}": {
      "get": {
        "operationId": "GetUsersForProject",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/slickqaUsersQueryResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "CompanyName",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "ProjectName",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/current": {
      "get": {
        "operationId": "GetCurrentUserInfo",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/slickqaUserInfo"
            }
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/info/{EmailAddress}": {
      "get": {
        "operationId": "GetUserInfo",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/slickqaUserInfo"
            }
          }
        },
        "parameters": [
          {
            "name": "EmailAddress",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
    "/api/version": {
      "get": {
        "operationId": "GetFullVersion",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/slickqaVersionInfoResponse"
            }
          }
        },
        "tags": [
          "Version"
        ]
      }
    }
  },
  "definitions": {
    "slickqaCompanyPermissionInfo": {
      "type": "object",
      "properties": {
        "CompanyName": {
          "type": "string"
        },
        "CompanyAdmin": {
          "type": "integer",
          "format": "int64"
        },
        "Projects": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/slickqaProjectPermissionInfo"
          }
        }
      }
    },
    "slickqaIsAuthorizedResponse": {
      "type": "object",
      "properties": {
        "Allowed": {
          "type": "boolean",
          "format": "boolean"
        },
        "Message": {
          "type": "string"
        }
      }
    },
    "slickqaLoginResponse": {
      "type": "object",
      "properties": {
        "Success": {
          "type": "boolean",
          "format": "boolean"
        },
        "Token": {
          "type": "string"
        },
        "User": {
          "$ref": "#/definitions/slickqaUserInfo"
        }
      }
    },
    "slickqaPlainUserLoginRequest": {
      "type": "object",
      "properties": {
        "UserName": {
          "type": "string"
        },
        "Password": {
          "type": "string"
        }
      }
    },
    "slickqaProjectPermissionInfo": {
      "type": "object",
      "properties": {
        "ProjectName": {
          "type": "string"
        },
        "Roles": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "slickqaSlickPermissionInfo": {
      "type": "object",
      "properties": {
        "SlickAdmin": {
          "type": "integer",
          "format": "int64"
        },
        "Companies": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/slickqaCompanyPermissionInfo"
          }
        }
      }
    },
    "slickqaUserInfo": {
      "type": "object",
      "properties": {
        "EmailAddress": {
          "type": "string"
        },
        "Permissions": {
          "$ref": "#/definitions/slickqaSlickPermissionInfo"
        },
        "FullName": {
          "type": "string"
        },
        "GivenName": {
          "type": "string"
        },
        "FamilyName": {
          "type": "string"
        },
        "AvatarUrl": {
          "type": "string"
        },
        "JobTitle": {
          "type": "string"
        },
        "HashedPassword": {
          "type": "string"
        }
      }
    },
    "slickqaUsersQueryResponse": {
      "type": "object",
      "properties": {
        "users": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/slickqaUserInfo"
          }
        }
      }
    },
    "slickqaVersionInfoResponse": {
      "type": "object",
      "properties": {
        "Version": {
          "type": "string"
        }
      }
    }
  }
}